LPS-53578 - update breadcrumb package
