+++ modules/apps/site-navigation/site-navigation-breadcrumb-web/src/META-INF/resources/init.jsp
@@ -21,7 +21,7 @@
 <%@ taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %>
 <%@ taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
 
-<%@ page import="com.liferay.breadcrumb.web.context.BreadcrumbDisplayContext" %><%@
+<%@ page import="com.liferay.site.navigation.breadcrumb.web.context.BreadcrumbDisplayContext" %><%@
 page import="com.liferay.portal.kernel.servlet.taglib.ui.BreadcrumbEntry" %><%@
 page import="com.liferay.portal.kernel.servlet.taglib.ui.BreadcrumbUtil" %><%@
 page import="com.liferay.portal.kernel.util.Constants" %><%@
