+++ modules/apps/site-navigation/site-navigation-breadcrumb-web/src/com/liferay/site/navigation/breadcrumb/web/portlet/template/dependencies/portlet_display_template_horizontal.ftl
